backend/requirements.txt
celery[redis]==5.2.7
dj-database-url==1.2.0 # connect to databases with url REVIEW: if not need
Django==4.1.5
django-celery-results==2.4.0
django-cors-headers==3.13.0 # cors header plugin
django-redis==5.2.0
django-storages==1.13.2 # Storage backend for django
djangorestframework==3.14.0
flower==1.2.0 # celery task visualizer
python-decouple==3.7 # Helper for django and environment variables
psutil==5.9.4 # process and system utilities
psycopg2-binary==2.9.5 # PostgreSQL database adapter for the Python
pandas==1.5.3 # data analysis and manipulation tool
scikit-learn==1.2.1 # machine learning in Python
